i2c: stm32: add support for STM32MP25 soc
The stm32mp25 embeds a new FMP bit in the CR1 register,
necessary when running at Fast Mode Plus frequency.
This bit should be used instead of the SYSCFG bit used on other platforms.

Add a new compatible "st,stm32mp25-i2c" to distinguish between the SoCs
and a boolean "fmp_cr1_bit" within the setup structure in order to know
if the platform has the FMP bit within CR1 is available or not.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I905b6a8eede51d56f7dee17a59b2efdee2c1dc7d
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/250312
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@foss.st.com>
Tested-by: Patrice CHOTARD <patrice.chotard@foss.st.com>