dm: adc: avoid overriding reference voltage values if already provided
Various drivers assume "VDD supplied by common vref pin" and this is
hardcoded in the framework. In case ADCs allow separate power supplies
for VDD (a.k.a. VDDA) and Vref, the driver must be allowed to specify the
value of vdd_microvolts/vss_microvolts, which is used by the u-class as
vref for raw to uV conversion , and avoid the u_class to override it after
that (when != -ENODATA).

Change-Id: Ic1da832ffa9d9813a154c83a218c5e9a824dc321
Signed-off-by: Cheick Traore <cheick.traore-ext@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/397459
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@foss.st.com>
Domain-Review: Fabrice GASNIER <fabrice.gasnier@foss.st.com>