board: stm32mp2: fix fwu_mdata node updating
Comparing the SD card boot with the eMMC one, the DTS differs as follows:

-       fwu-mdata-store = <0x00000085>;
+       fwu-mdata-store = "/soc@0/bus@42080000/mmc@48230000";

In both cases, the U-Boot DTS loaded by tf-a-stm32mp at address 0x84400000
was obtained with the following commands:

STM32MP> fdt addr 0x84400000
Working FDT set to 84400000
STM32MP> fdt print

As shown in the diff, the working case (SD card) uses a numeric value,
i.e. the phandle of the sdmmc1 node, while the eMMC case uses a path.
This patch updates the fwu-mdata-store property to use the phandle
instead of the path.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>